Replace ts-node with tsx (#20)
When the Prisma client is generated in a custom output path, the created
client does not work in scripts. This is likely due to ESM vs CommonJS
interop and runtime mismatches in `ts-node` scripts. Prisma generates
ESM-compatible code by default (even though it's technically CommonJS),
and `ts-node` can improperly resolve and instantiate it.